Bias looping to reducing text that's too tall
When hashing the sizes of the selected nodes, only use absoluteRenderBounds if the node is in a frame and stretched to its parent size.
Don't emit settingsChanged from the UI when the component is first mounting, since those settings weren't changed by the user.
Remove main field from package.json.